refactor: move models
