--- go-web/internal/Seller/models/models.go
@@ -1,9 +1,0 @@
-package models
-
-type Seller struct {
-	ID    int  `json:"id"`
-	Cid  string  `json:"cid"`
-	Company_name string  `json:"company_name"`
-	Address string  `json:"address"`
-	Telephone  string `json:"telephone"`
-}
