+++ go-web/internal/models/models.go
@@ -42,3 +42,11 @@ type Warehouse struct {
 	MinimunCapacity    int    `json:"minimun_capacity"`
 	MinimunTemperature int    `json:"minimun_temperature"`
 }
+
+type Seller struct {
+	ID    int  `json:"id"`
+	Cid  string  `json:"cid"`
+	Company_name string  `json:"company_name"`
+	Address string  `json:"address"`
+	Telephone  string `json:"telephone"`
+}
